requirements.txt
pyautogen[retrievechat]~=0.2.0b5
qdrant_client[fastembed]